Split the WPF component between Control and VideoSourceProvider
