+++ src/Samples/Vlc.DotNet.Wpf.Samples/MainWindow.xaml.cs
@@ -41,16 +41,16 @@ namespace Vlc.DotNet.Wpf.Samples
             this.control?.Dispose();
             this.control = new VlcControl();
             this.ControlContainer.Content = this.control;
-            this.control.CreatePlayer(this.vlcLibDirectory);
+            this.control.SourceProvider.CreatePlayer(this.vlcLibDirectory);
 
             // This can also be called before EndInit
-            this.control.MediaPlayer.Log += (_, args) =>
+            this.control.SourceProvider.MediaPlayer.Log += (_, args) =>
             {
                 string message = $"libVlc : {args.Level} {args.Message} @ {args.Module}";
                 System.Diagnostics.Debug.WriteLine(message);
             };
 
-            control.MediaPlayer.Play(new Uri("http://download.blender.org/peach/bigbuckbunny_movies/big_buck_bunny_480p_surround-fix.avi"));
+            control.SourceProvider.MediaPlayer.Play(new Uri("http://download.blender.org/peach/bigbuckbunny_movies/big_buck_bunny_480p_surround-fix.avi"));
             //control.MediaPlayer.Play(new FileInfo(@"..\..\..\Vlc.DotNet\Samples\Videos\BBB trailer.mov"));
         }
 
@@ -67,7 +67,7 @@ namespace Vlc.DotNet.Wpf.Samples
                 return;
             }
 
-            this.control.MediaPlayer.Rate = 2;
+            this.control.SourceProvider.MediaPlayer.Rate = 2;
         }
 
         private void GetLength_Click(object sender, RoutedEventArgs e)
@@ -77,7 +77,7 @@ namespace Vlc.DotNet.Wpf.Samples
                 return;
             }
 
-            GetLength.Content = this.control.MediaPlayer.Length + " ms";
+            GetLength.Content = this.control.SourceProvider.MediaPlayer.Length + " ms";
         }
 
         private void GetCurrentTime_Click(object sender, RoutedEventArgs e)
@@ -87,7 +87,7 @@ namespace Vlc.DotNet.Wpf.Samples
                 return;
             }
 
-            GetCurrentTime.Content = this.control.MediaPlayer.Time + " ms";
+            GetCurrentTime.Content = this.control.SourceProvider.MediaPlayer.Time + " ms";
         }
 
         private void SetCurrentTime_Click(object sender, RoutedEventArgs e)
@@ -97,8 +97,8 @@ namespace Vlc.DotNet.Wpf.Samples
                 return;
             }
 
-            this.control.MediaPlayer.Time = 5000;
-            SetCurrentTime.Content = this.control.MediaPlayer.Time + " ms";
+            this.control.SourceProvider.MediaPlayer.Time = 5000;
+            SetCurrentTime.Content = this.control.SourceProvider.MediaPlayer.Time + " ms";
         }
     }
 }
